+++ tests/test_prompt_template.py
@@ -44,6 +44,15 @@ def test_title4():
         template.update_title(25)
 
 
+def test_title5():
+    template = PromptTemplate(
+        content="Act as a {language} developer and respond to this question:\n{prompt_message}",
+        custom_map={
+            "language": "Python"},
+        title="")
+    assert template.title == ""
+
+
 def test_content1():
     template = PromptTemplate(
         content="Act as a {language} developer and respond to this question:\n{prompt_message}",
@@ -70,6 +79,14 @@ def test_content3():
         template.update_content(content=22)
 
 
+def test_content4():
+    template = PromptTemplate(
+        content="",
+        custom_map={
+            "language": "Python"})
+    assert template.content == ""
+
+
 def test_custom_map1():
     template = PromptTemplate(
         content="Act as a {language} developer and respond to this question:\n{prompt_message}",
@@ -94,6 +111,13 @@ def test_custom_map3():
             "language": "Python"})
     with pytest.raises(MemorValidationError, match=r"Invalid custom map: it must be a dictionary with keys and values that can be converted to strings."):
         template.update_map(["C++"])
+
+
+def test_custom_map4():
+    template = PromptTemplate(
+        content="Act as a {language} developer and respond to this question:\n{prompt_message}",
+        custom_map={})
+    assert template.custom_map == {}
 
 
 def test_date_modified():