+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `AI_STUDIO` render format modified
+- Test system modified
 ## [0.7] - 2025-06-25
 ### Added
 - `Message` abstract class